@@ -6,7 +6,7 @@
     <parent>
         <groupId>org.springframework.boot</groupId>
         <artifactId>spring-boot-starter-parent</artifactId>
-        <version>3.5.4</version>
+        <version>4.0.1</version>
         <relativePath/> <!-- lookup parent from repository -->
     </parent>
     <groupId>org.ost.advertisement</groupId>
@@ -16,8 +16,8 @@
     <description>Project for Advertisement</description>
 
     <properties>
-        <java.version>21</java.version>
-        <vaadin.version>24.9.4</vaadin.version>
+        <java.version>25</java.version>
+        <vaadin.version>25.0.4</vaadin.version>
         <mapstruct.version>1.6.3</mapstruct.version>
         <lombok-mapstruct-binding.version>0.2.0</lombok-mapstruct-binding.version>
     </properties>
@@ -45,6 +45,11 @@
         <dependency>
           <groupId>org.springframework.boot</groupId>
           <artifactId>spring-boot-starter-security</artifactId>
+        </dependency>
+<!-- Validation -->
+        <dependency>
+            <groupId>org.springframework.boot</groupId>
+            <artifactId>spring-boot-starter-validation</artifactId>
         </dependency>
 <!-- PostgresSQL JDBC Driver -->
         <dependency>
@@ -79,6 +84,24 @@
           <version>${lombok-mapstruct-binding.version}</version>
           <scope>provided</scope>
         </dependency>
+
+<!-- Live reload -->
+        <dependency>
+            <groupId>com.vaadin</groupId>
+            <artifactId>vaadin-dev</artifactId>
+            <version>${vaadin.version}</version>
+            <scope>runtime</scope>
+            <optional>true</optional>
+        </dependency>
+
+<!-- Compile‑time only -->
+        <dependency>
+            <groupId>org.jetbrains</groupId>
+            <artifactId>annotations</artifactId>
+            <version>24.1.0</version>
+            <scope>provided</scope>
+        </dependency>
+
 <!--for test part-->
         <dependency>
             <groupId>org.springframework.boot</groupId>
@@ -131,6 +154,7 @@
         <configuration>
           <source>${java.version}</source>
           <target>${java.version}</target>
+          <encoding>UTF-8</encoding>
           <annotationProcessorPaths>
             <path>
               <groupId>org.mapstruct</groupId>
