@@ -2,40 +2,74 @@ package org.ost.advertisement.configuration;
 
 import com.vaadin.flow.server.HandlerHelper;
 import jakarta.servlet.http.HttpServletRequest;
+import lombok.RequiredArgsConstructor;
+import org.ost.advertisement.repository.user.UserRepository;
+import org.ost.advertisement.security.UserPrincipal;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.security.config.Customizer;
+import org.springframework.security.authentication.AuthenticationManager;
+import org.springframework.security.config.annotation.authentication.configuration.AuthenticationConfiguration;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
 import org.springframework.security.config.http.SessionCreationPolicy;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.security.web.SecurityFilterChain;
+import org.springframework.security.web.context.HttpSessionSecurityContextRepository;
+import org.springframework.security.web.context.SecurityContextRepository;
+import org.springframework.security.web.util.matcher.RequestMatcher;
 
 @Configuration
+@RequiredArgsConstructor
 public class SecurityConfig {
 
-	private static final String SERVLET_MAPPING_PATH = "/";
+    private static final String SERVLET_MAPPING_PATH = "/";
 
-	@Bean
-	public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
-		http
-			.authorizeHttpRequests(auth -> auth
-				.requestMatchers(this::isVaadinInternalRequest).permitAll()
-				.anyRequest().permitAll()
-			)
-			.csrf(csrf -> csrf
-				.ignoringRequestMatchers(this::isVaadinInternalRequest)
-			)
-			.sessionManagement(session -> session
-				.sessionCreationPolicy(SessionCreationPolicy.IF_REQUIRED)
-			)
-			.httpBasic(Customizer.withDefaults())
-			.formLogin(AbstractHttpConfigurer::disable)
-			.logout(AbstractHttpConfigurer::disable);
+    private final RequestMatcher vaadinInternalRequestMatcher = this::isVaadinInternalRequest;
+    private final UserRepository userRepository;
 
-		return http.build();
-	}
+    @Bean
+    public PasswordEncoder passwordEncoder() {
+        return new BCryptPasswordEncoder();
+    }
 
-	private boolean isVaadinInternalRequest(HttpServletRequest request) {
-		return HandlerHelper.isFrameworkInternalRequest(SERVLET_MAPPING_PATH, request);
-	}
+    @Bean
+    public UserDetailsService userDetailsService() {
+        return email -> userRepository.findByEmail(email)
+                .map(UserPrincipal::new)
+                .orElseThrow(() -> new UsernameNotFoundException("User not found: " + email));
+    }
+
+    @Bean
+    public AuthenticationManager authenticationManager(AuthenticationConfiguration config) throws Exception {
+        return config.getAuthenticationManager();
+    }
+
+    @Bean
+    public SecurityContextRepository securityContextRepository() {
+        return new HttpSessionSecurityContextRepository();
+    }
+
+    @Bean
+    public SecurityFilterChain securityFilterChain(HttpSecurity http,
+                                                   SecurityContextRepository securityContextRepository) throws Exception {
+        http
+                .authorizeHttpRequests(auth -> auth
+                        .requestMatchers(vaadinInternalRequestMatcher).permitAll()
+                        .anyRequest().permitAll()
+                )
+                .securityContext(ctx -> ctx.securityContextRepository(securityContextRepository))
+                .csrf(csrf -> csrf.ignoringRequestMatchers(vaadinInternalRequestMatcher))
+                .sessionManagement(session -> session.sessionCreationPolicy(SessionCreationPolicy.IF_REQUIRED))
+                .formLogin(AbstractHttpConfigurer::disable)
+                .logout(AbstractHttpConfigurer::disable);
+
+        return http.build();
+    }
+
+    private boolean isVaadinInternalRequest(HttpServletRequest request) {
+        return HandlerHelper.isFrameworkInternalRequest(SERVLET_MAPPING_PATH, request);
+    }
 }
