@@ -1,16 +1,25 @@
 package org.ost.advertisement.configuration.db;
 
-import org.springframework.context.annotation.Bean;
+import lombok.RequiredArgsConstructor;
+import org.jspecify.annotations.NonNull;
+import org.ost.advertisement.entities.User;
+import org.ost.advertisement.services.auth.AuthContextService;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.domain.AuditorAware;
 import org.springframework.data.jdbc.repository.config.EnableJdbcAuditing;
 
-@Configuration
-@EnableJdbcAuditing(auditorAwareRef = "auditorProvider")
-public class AuditingConfig {
+import java.util.Optional;
 
-	@Bean
-	public AuditorAware<Long> auditorProvider() {
-		return new SecurityAuditorAware();
-	}
+@Configuration
+@EnableJdbcAuditing
+@RequiredArgsConstructor
+public class AuditingConfig implements AuditorAware<Long> {
+
+    private final AuthContextService authContextService;
+
+    @Override
+    public @NonNull Optional<Long> getCurrentAuditor() {
+        return authContextService.getCurrentUser().map(User::getId);
+    }
 }
+
