@@ -2,7 +2,6 @@ package org.ost.advertisement.dto.filter;
 
 import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.Size;
-import java.time.Instant;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -10,6 +9,8 @@ import lombok.NoArgsConstructor;
 import lombok.experimental.FieldNameConstants;
 import org.ost.advertisement.entities.Role;
 import org.ost.advertisement.validation.ValidRange;
+
+import java.time.Instant;
 
 @FieldNameConstants
 @Data
@@ -21,23 +22,23 @@ import org.ost.advertisement.validation.ValidRange;
 @ValidRange(start = "startId", end = "endId", message = "startId must not be greater than endId")
 public class UserFilterDto {
 
-	@Size(max = 255, message = "Name must not exceed 255 characters")
-	private String name;
-	@Size(max = 255, message = "Name must not exceed 255 characters")
-	private String email;
-	private Role role;
-	private Instant createdAtStart;
-	private Instant createdAtEnd;
-	private Instant updatedAtStart;
-	private Instant updatedAtEnd;
+    @Size(max = 255, message = "Name must not exceed 255 characters")
+    private String name;
+    @Size(max = 255, message = "Email must not exceed 255 characters")
+    private String email;
+    private Role role;
+    private Instant createdAtStart;
+    private Instant createdAtEnd;
+    private Instant updatedAtStart;
+    private Instant updatedAtEnd;
 
-	@Min(value = 0, message = "Start ID must be non-negative (>= {value}).")
-	private Long startId;
-	@Min(value = 0, message = "End ID must be non-negative (>= {value}).")
-	private Long endId;
+    @Min(value = 0, message = "Start ID must be non-negative (>= {value}).")
+    private Long startId;
+    @Min(value = 0, message = "End ID must be non-negative (>= {value}).")
+    private Long endId;
 
-	public static UserFilterDto empty() {
-		return new UserFilterDto();
-	}
+    public static UserFilterDto empty() {
+        return new UserFilterDto();
+    }
 
 }
