@@ -1,7 +1,7 @@
 package org.ost.advertisement.entities;
 
 public enum Role {
-	USER,
-	ADMIN,
-	MODERATOR
+    USER,
+    ADMIN,
+    MODERATOR
 }
