@@ -2,8 +2,8 @@ package org.ost.advertisement.exceptions.authorization;
 
 public class AccessDeniedException extends AuthorizationException {
 
-	public AccessDeniedException(String message) {
-		super(message);
-	}
+    public AccessDeniedException(String message) {
+        super(message);
+    }
 }
 
