@@ -1,7 +1,7 @@
 package org.ost.advertisement.exceptions.authorization;
 
-public abstract class AuthorizationException extends RuntimeException {
-	public AuthorizationException(String message) {
-		super(message);
-	}
+abstract class AuthorizationException extends RuntimeException {
+    protected AuthorizationException(String message) {
+        super(message);
+    }
 }
