@@ -7,7 +7,7 @@ import org.ost.advertisement.dto.filter.AdvertisementFilterDto;
 @Mapper(componentModel = "spring")
 public interface AdvertisementFilterMapper extends FilterMapper<AdvertisementFilterDto> {
 
-	void update(@MappingTarget AdvertisementFilterDto target, AdvertisementFilterDto source);
+    void update(@MappingTarget AdvertisementFilterDto target, AdvertisementFilterDto source);
 
-	AdvertisementFilterDto copy(AdvertisementFilterDto source);
+    AdvertisementFilterDto copy(AdvertisementFilterDto source);
 }
