@@ -8,13 +8,13 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 public class AdvertisementRepositoryCustomImpl
-	extends RepositoryCustom<AdvertisementInfoDto, AdvertisementFilterDto>
-	implements AdvertisementRepositoryCustom {
+        extends RepositoryCustom<AdvertisementInfoDto, AdvertisementFilterDto>
+        implements AdvertisementRepositoryCustom {
 
-	private static final AdvertisementProjection ADVERTISEMENT_PROJECTION = new AdvertisementProjection();
-	private static final AdvertisementFilterBuilder ADVERTISEMENT_FILTER_BUILDER = new AdvertisementFilterBuilder();
+    private static final AdvertisementProjection ADVERTISEMENT_PROJECTION = new AdvertisementProjection();
+    private static final AdvertisementFilterBuilder ADVERTISEMENT_FILTER_BUILDER = new AdvertisementFilterBuilder();
 
-	public AdvertisementRepositoryCustomImpl(NamedParameterJdbcTemplate jdbc) {
-		super(jdbc, ADVERTISEMENT_PROJECTION, ADVERTISEMENT_FILTER_BUILDER);
-	}
+    public AdvertisementRepositoryCustomImpl(NamedParameterJdbcTemplate jdbc) {
+        super(jdbc, ADVERTISEMENT_PROJECTION, ADVERTISEMENT_FILTER_BUILDER);
+    }
 }
