@@ -2,7 +2,7 @@ package org.ost.advertisement.repository.query.filter;
 
 public interface FilterMapping {
 
-	String filterProperty();
+    String filterProperty();
 
-	String sqlExpression();
+    String sqlExpression();
 }
