@@ -2,7 +2,7 @@ package org.ost.advertisement.repository.query.projection;
 
 public interface SqlFieldProjection {
 
-	String sqlExpression();
+    String sqlExpression();
 
-	String alias();
+    String alias();
 }
