@@ -2,6 +2,6 @@ package org.ost.advertisement.security;
 
 public interface UserIdMarker {
 
-	Long getOwnerUserId();
+    Long getOwnerUserId();
 
 }
