@@ -1,11 +1,8 @@
 package org.ost.advertisement.ui.dto;
 
+import lombok.*;
+
 import java.time.Instant;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
 
 @NoArgsConstructor
 @AllArgsConstructor
@@ -14,14 +11,14 @@ import lombok.Setter;
 @Setter
 public class AdvertisementEditDto {
 
-	private Long id;
+    private Long id;
 
-	private String title;
-	private String description;
+    private String title;
+    private String description;
 
-	private Instant createdAt;
-	private Instant updatedAt;
+    private Instant createdAt;
+    private Instant updatedAt;
 
-	private Long createdByUserId;
-	private Long lastModifiedByUserId;
+    private Long createdByUserId;
+    private Long lastModifiedByUserId;
 }
