@@ -3,16 +3,16 @@ package org.ost.advertisement.ui.mappers;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
 import org.ost.advertisement.dto.AdvertisementInfoDto;
-import org.ost.advertisement.ui.dto.AdvertisementEditDto;
 import org.ost.advertisement.entities.Advertisement;
+import org.ost.advertisement.ui.dto.AdvertisementEditDto;
 
 @Mapper(componentModel = "spring")
 public interface AdvertisementMapper {
 
-	@Mapping(target = "updatedAt", ignore = true)
-	@Mapping(target = "lastModifiedByUserId", ignore = true)
-	Advertisement toAdvertisement(AdvertisementEditDto dto);
+    @Mapping(target = "updatedAt", ignore = true)
+    @Mapping(target = "lastModifiedByUserId", ignore = true)
+    Advertisement toAdvertisement(AdvertisementEditDto dto);
 
-	AdvertisementEditDto toAdvertisementEdit(AdvertisementInfoDto dto);
+    AdvertisementEditDto toAdvertisementEdit(AdvertisementInfoDto dto);
 
 }
