@@ -4,17 +4,17 @@ import com.vaadin.flow.component.notification.Notification;
 import com.vaadin.flow.component.notification.NotificationVariant;
 
 public enum NotificationType {
-	SUCCESS(NotificationVariant.LUMO_SUCCESS),
-	ERROR(NotificationVariant.LUMO_ERROR);
+    SUCCESS(NotificationVariant.LUMO_SUCCESS),
+    ERROR(NotificationVariant.LUMO_ERROR);
 
-	private final NotificationVariant variant;
+    private final NotificationVariant variant;
 
-	NotificationType(NotificationVariant variant) {
-		this.variant = variant;
-	}
+    NotificationType(NotificationVariant variant) {
+        this.variant = variant;
+    }
 
-	public void show(String text) {
-		Notification notification = Notification.show(text, 4000, Notification.Position.BOTTOM_START);
-		notification.addThemeVariants(variant);
-	}
+    public void show(String text) {
+        Notification notification = Notification.show(text, 4000, Notification.Position.BOTTOM_START);
+        notification.addThemeVariants(variant);
+    }
 }
