@@ -5,23 +5,17 @@ import org.ost.advertisement.ui.utils.SvgUtils;
 
 public class SvgIcon extends Span {
 
-	public SvgIcon(String resourcePath) {
-		loadSvg(resourcePath);
-		getStyle().set("width", "1em");
-		getStyle().set("height", "1em");
-		getStyle().set("display", "inline-block");
-		getStyle().set("vertical-align", "middle");
-		getStyle().set("line-height", "1");
-	}
+    public SvgIcon(String resourcePath) {
+        addClassName("svg-icon");
+        setSvg(resourcePath);
+    }
 
-	public void loadSvg(String resourcePath) {
-		String svg = SvgUtils.loadSvg(resourcePath);
-		getElement().setProperty("innerHTML", svg);
-	}
-
-	public void setColor(String color) {
-		this.getStyle().set("color", color);
-	}
-
+    public void setSvg(String resourcePath) {
+        String svg = SvgUtils.loadSvg(resourcePath);
+        if (!svg.isBlank()) {
+            getElement().setProperty("innerHTML", svg);
+        }
+    }
 }
+
 
