@@ -1,7 +1,6 @@
 package org.ost.advertisement.ui.views.users;
 
 import com.vaadin.flow.component.grid.Grid;
-import com.vaadin.flow.component.html.Div;
 import com.vaadin.flow.component.orderedlayout.VerticalLayout;
 import lombok.Getter;
 import org.ost.advertisement.entities.User;
@@ -11,27 +10,29 @@ import org.ost.advertisement.ui.views.components.PaginationBarModern;
 @Getter
 public class UserLayout extends VerticalLayout {
 
-	private final Grid<User> grid = new Grid<>(User.class, false);
-	private final PaginationBarModern paginationBar;
+    private final Grid<User> grid;
+    private final PaginationBarModern paginationBar;
 
-	public UserLayout(I18nService i18n) {
-		this.paginationBar = new PaginationBarModern(i18n);
+    public UserLayout(I18nService i18n) {
+        this.grid = new Grid<>(User.class, false);
+        this.paginationBar = new PaginationBarModern(i18n);
 
-		addClassName("user-list-layout");
-		setSizeFull();
-		setPadding(false);
-		setSpacing(false);
+        initLayout();
+    }
 
-		grid.setSizeFull();
+    private void initLayout() {
+        // CSS class names only; visual rules moved to CSS
+        addClassName("user-list-layout");
 
-		Div gridWrapper = new Div(grid);
-		gridWrapper.setWidthFull();
-		gridWrapper.setHeight("100%");
-		gridWrapper.getStyle().set("overflow-x", "auto");
-		gridWrapper.getStyle().set("overflow-y", "auto");
-		gridWrapper.getStyle().set("flex-grow", "1");
+        // keep width behavior in code or CSS (both ok)
+        setWidthFull();
 
-		setFlexGrow(1, gridWrapper);
-		add(gridWrapper, paginationBar);
-	}
+        grid.addClassName("user-grid");
+        grid.setWidthFull();
+        grid.setAllRowsVisible(true);
+
+        paginationBar.addClassName("user-pagination");
+
+        add(grid, paginationBar);
+    }
 }
