@@ -1,165 +1,124 @@
 package org.ost.advertisement.ui.views.users.dialogs;
 
-import static java.util.Optional.ofNullable;
-import static org.ost.advertisement.constants.I18nKey.USER_DIALOG_BUTTON_CANCEL;
-import static org.ost.advertisement.constants.I18nKey.USER_DIALOG_BUTTON_SAVE;
-import static org.ost.advertisement.constants.I18nKey.USER_DIALOG_FIELD_CREATED_LABEL;
-import static org.ost.advertisement.constants.I18nKey.USER_DIALOG_FIELD_EMAIL_LABEL;
-import static org.ost.advertisement.constants.I18nKey.USER_DIALOG_FIELD_ID_LABEL;
-import static org.ost.advertisement.constants.I18nKey.USER_DIALOG_FIELD_NAME_LABEL;
-import static org.ost.advertisement.constants.I18nKey.USER_DIALOG_FIELD_NAME_PLACEHOLDER;
-import static org.ost.advertisement.constants.I18nKey.USER_DIALOG_FIELD_ROLE_LABEL;
-import static org.ost.advertisement.constants.I18nKey.USER_DIALOG_FIELD_UPDATED_LABEL;
-import static org.ost.advertisement.constants.I18nKey.USER_DIALOG_NOTIFICATION_SAVE_ERROR;
-import static org.ost.advertisement.constants.I18nKey.USER_DIALOG_NOTIFICATION_SUCCESS;
-import static org.ost.advertisement.constants.I18nKey.USER_DIALOG_TITLE;
-import static org.ost.advertisement.constants.I18nKey.USER_DIALOG_VALIDATION_NAME_LENGTH;
-import static org.ost.advertisement.constants.I18nKey.USER_DIALOG_VALIDATION_NAME_REQUIRED;
-import static org.ost.advertisement.constants.I18nKey.USER_DIALOG_VALIDATION_ROLE_REQUIRED;
-import static org.ost.advertisement.ui.utils.TimeZoneUtil.formatInstant;
-import static org.ost.advertisement.ui.views.TailwindStyle.EMAIL_LABEL;
-import static org.ost.advertisement.ui.views.TailwindStyle.GRAY_LABEL;
-
-import com.vaadin.flow.component.button.Button;
-import com.vaadin.flow.component.combobox.ComboBox;
-import com.vaadin.flow.component.textfield.TextField;
 import com.vaadin.flow.data.validator.StringLengthValidator;
 import com.vaadin.flow.spring.annotation.SpringComponent;
-import java.util.Arrays;
-import java.util.Objects;
-import lombok.Getter;
+import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.ost.advertisement.entities.Role;
 import org.ost.advertisement.entities.User;
 import org.ost.advertisement.services.I18nService;
 import org.ost.advertisement.services.UserService;
 import org.ost.advertisement.ui.dto.UserEditDto;
 import org.ost.advertisement.ui.mappers.UserMapper;
-import org.ost.advertisement.ui.views.components.dialogs.DialogContentFactory;
 import org.ost.advertisement.ui.views.components.dialogs.FormDialogDelegate;
-import org.ost.advertisement.ui.views.components.dialogs.LabeledField;
+import org.ost.advertisement.ui.views.users.dialogs.fields.*;
 import org.springframework.beans.factory.ObjectProvider;
 import org.springframework.context.annotation.Scope;
+
+import java.util.Objects;
+
+import static java.util.Optional.ofNullable;
+import static org.ost.advertisement.constants.I18nKey.*;
+import static org.ost.advertisement.ui.utils.TimeZoneUtil.formatInstant;
 
 @SpringComponent
 @Scope("prototype")
 @Slf4j
+@RequiredArgsConstructor
 public class UserEditDialog {
 
-	private final UserService userService;
-	private final UserMapper mapper;
-	private final LabeledField.Builder labeledFieldBuilder;
-	private final I18nService i18n;
-	@Getter
-	private final FormDialogDelegate<UserEditDto> delegate;
+    private final UserService userService;
+    private final UserMapper mapper;
+    private final I18nService i18n;
+    private final DialogUserIdLabeledField idField;
+    private final DialogUserEmailLabeledField emailField;
+    private final DialogUserCreatedAtLabeledField createdAtField;
+    private final DialogUserUpdatedAtLabeledField updatedAtField;
+    private final DialogUserNameTextField nameField;
+    private final DialogUserRoleComboBox roleCombo;
+    private final DialogUserSaveButton saveButton;
+    private final DialogUserCancelButton cancelButton;
 
-	private UserEditDialog(UserService userService,
-						   UserMapper mapper,
-						   LabeledField.Builder labeledFieldBuilder,
-						   I18nService i18n,
-						   FormDialogDelegate<UserEditDto> delegate) {
-		this.userService = userService;
-		this.mapper = mapper;
-		this.labeledFieldBuilder = labeledFieldBuilder;
-		this.i18n = i18n;
-		this.delegate = delegate;
-	}
+    private FormDialogDelegate<UserEditDto> delegate;
 
-	private void configureDialog() {
-		UserEditDto user = delegate.getDto();
-		delegate.setTitle(i18n.get(USER_DIALOG_TITLE));
+    private void configureDialog(FormDialogDelegate<UserEditDto> delegate) {
+        this.delegate = delegate;
+        setTitle();
+        bindFields();
+        updateMetadata();
+        addContent();
+        addActions();
+    }
 
-		TextField nameField = DialogContentFactory.textField(
-			i18n, USER_DIALOG_FIELD_NAME_LABEL, USER_DIALOG_FIELD_NAME_PLACEHOLDER, 255, true
-		);
+    private void setTitle() {
+        delegate.setTitle(i18n.get(USER_DIALOG_TITLE));
+    }
 
-		ComboBox<Role> roleCombo = DialogContentFactory.comboBox(
-			i18n, USER_DIALOG_FIELD_ROLE_LABEL, Arrays.asList(Role.values()), true
-		);
+    private void bindFields() {
+        delegate.getBinder().forField(nameField)
+                .asRequired(i18n.get(USER_DIALOG_VALIDATION_NAME_REQUIRED))
+                .withValidator(new StringLengthValidator(i18n.get(USER_DIALOG_VALIDATION_NAME_LENGTH), 1, 255))
+                .bind(UserEditDto::getName, UserEditDto::setName);
 
-		delegate.getBinder().forField(nameField)
-			.asRequired(i18n.get(USER_DIALOG_VALIDATION_NAME_REQUIRED))
-			.withValidator(new StringLengthValidator(i18n.get(USER_DIALOG_VALIDATION_NAME_LENGTH), 1, 255))
-			.bind(UserEditDto::getName, UserEditDto::setName);
+        delegate.getBinder().forField(roleCombo)
+                .asRequired(i18n.get(USER_DIALOG_VALIDATION_ROLE_REQUIRED))
+                .bind(UserEditDto::getRole, UserEditDto::setRole);
+    }
 
-		delegate.getBinder().forField(roleCombo)
-			.asRequired(i18n.get(USER_DIALOG_VALIDATION_ROLE_REQUIRED))
-			.bind(UserEditDto::getRole, UserEditDto::setRole);
+    private void updateMetadata() {
+        UserEditDto user = delegate.getDto();
+        idField.update(String.valueOf(user.getId()));
+        emailField.update(ofNullable(user.getEmail()).orElse(""));
+        createdAtField.update(formatInstant(user.getCreatedAt()));
+        updatedAtField.update(formatInstant(user.getUpdatedAt()));
+    }
 
-		delegate.addContent(
-			labeledFieldBuilder.withLabel(USER_DIALOG_FIELD_ID_LABEL)
-				.withValue(String.valueOf(user.getId()))
-				.withStyles(EMAIL_LABEL)
-				.build(),
-			labeledFieldBuilder.withLabel(USER_DIALOG_FIELD_EMAIL_LABEL)
-				.withValue(ofNullable(user.getEmail()).orElse(""))
-				.withStyles(EMAIL_LABEL)
-				.build(),
-			nameField,
-			roleCombo,
-			labeledFieldBuilder.withLabel(USER_DIALOG_FIELD_CREATED_LABEL)
-				.withValue(formatInstant(user.getCreatedAt()))
-				.withStyles(GRAY_LABEL)
-				.build(),
-			labeledFieldBuilder.withLabel(USER_DIALOG_FIELD_UPDATED_LABEL)
-				.withValue(formatInstant(user.getUpdatedAt()))
-				.withStyles(GRAY_LABEL)
-				.build()
-		);
+    private void addContent() {
+        delegate.addContent(idField, emailField, nameField, roleCombo, createdAtField, updatedAtField);
+    }
 
-		Button saveButton = DialogContentFactory.primaryButton(i18n, USER_DIALOG_BUTTON_SAVE);
-		saveButton.addClickListener(event -> delegate.save(
-			u -> userService.save(mapper.toUser(u)),
-			USER_DIALOG_NOTIFICATION_SUCCESS,
-			USER_DIALOG_NOTIFICATION_SAVE_ERROR
-		));
+    private void addActions() {
+        saveButton.addClickListener(_ -> delegate.save(
+                u -> userService.save(mapper.toUser(u)),
+                USER_DIALOG_NOTIFICATION_SUCCESS,
+                USER_DIALOG_NOTIFICATION_SAVE_ERROR
+        ));
 
-		Button cancelButton = DialogContentFactory.tertiaryButton(i18n, USER_DIALOG_BUTTON_CANCEL);
-		cancelButton.addClickListener(event -> delegate.close());
+        cancelButton.addClickListener(_ -> delegate.close());
 
-		delegate.addActions(saveButton, cancelButton);
-	}
+        delegate.addActions(saveButton, cancelButton);
+    }
 
-	@SpringComponent
-	public static class Builder {
+    public void open() {
+        delegate.open();
+    }
 
-		private final UserService userService;
-		private final UserMapper mapper;
-		private final LabeledField.Builder labeledFieldBuilder;
-		private final I18nService i18n;
-		private final FormDialogDelegate.Builder<UserEditDto> delegateBuilder;
-		private final ObjectProvider<UserEditDialog> dialogProvider;
+    @SpringComponent
+    @RequiredArgsConstructor
+    public static class Builder {
 
-		public Builder(UserService userService,
-					   UserMapper mapper,
-					   LabeledField.Builder labeledFieldBuilder,
-					   I18nService i18n,
-					   FormDialogDelegate.Builder<UserEditDto> delegateBuilder,
-					   ObjectProvider<UserEditDialog> dialogProvider) {
-			this.userService = userService;
-			this.mapper = mapper;
-			this.labeledFieldBuilder = labeledFieldBuilder;
-			this.i18n = i18n;
-			this.delegateBuilder = delegateBuilder;
-			this.dialogProvider = dialogProvider;
-		}
+        private final UserMapper mapper;
+        private final FormDialogDelegate.Builder<UserEditDto> delegateBuilder;
+        private final ObjectProvider<UserEditDialog> dialogProvider;
 
-		public UserEditDialog build(User user, Runnable refresh) {
-			FormDialogDelegate<UserEditDto> delegate = delegateBuilder
-				.withClass(UserEditDto.class)
-				.withDto(mapper.toUserEdit(Objects.requireNonNull(user)))
-				.withRefresh(refresh)
-				.build();
+        public UserEditDialog build(User user, Runnable refresh) {
+            FormDialogDelegate<UserEditDto> delegate = createDelegate(user, refresh);
+            UserEditDialog dialog = dialogProvider.getObject();
+            dialog.configureDialog(delegate);
+            return dialog;
+        }
 
-			UserEditDialog dialog = dialogProvider.getObject(userService, mapper, labeledFieldBuilder, i18n, delegate);
-			dialog.configureDialog();
-			return dialog;
-		}
+        public UserEditDialog buildAndOpen(User user, Runnable refresh) {
+            UserEditDialog dialog = build(user, refresh);
+            dialog.open();
+            return dialog;
+        }
 
-		public UserEditDialog buildAndOpen(User user, Runnable refresh) {
-			UserEditDialog dialog = build(user, refresh);
-			dialog.getDelegate().open();
-			return dialog;
-		}
-	}
+        private FormDialogDelegate<UserEditDto> createDelegate(User user, Runnable refresh) {
+            return delegateBuilder
+                    .withClass(UserEditDto.class)
+                    .withDto(mapper.toUserEdit(Objects.requireNonNull(user)))
+                    .withRefresh(refresh)
+                    .build();
+        }
+    }
 }
